move svgs and icons to own asset folders
