+++ frontend/src/Assets/Icons/StarIcon.jsx
