+++ frontend/src/Assets/SVGs/ConfirmResetSvg.jsx
