+++ frontend/src/Assets/SVGs/EmailSvg.jsx
