+++ frontend/src/Assets/SVGs/ForgotPWSvg.jsx
