+++ frontend/src/Assets/SVGs/HeroSvg.jsx
