+++ frontend/src/Assets/SVGs/RegisterSvg.jsx
